back/requirements.txt
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.11
iniconfig==2.3.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.5
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
views==0.3

python-stdnum~=2.1
django-filter~=25.2
stripe~=13.1.2
weasyprint~=66.0
celery[redis]
redis
django-celery-beat
celery
python-dateutil~=2.9.0.post0
geopy~=2.4.1
channels
channels-redis
bleach~=6.3.0
webencodings~=0.5.1
daphne